size of various types

+++ 0x00-hello_world/6-size.c
@@ -13,8 +13,8 @@ int main(void)
 
 
 	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %lu byte(s)\n", (unsigned int)sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long int)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long long)sizeof(c));
 	printf("Size of long long int: %lu byte(s)\n", (unsigned long long)sizeof(d));
 	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
